+++ components/Coaching/CoachingCompanies/CoachingCompanies.controller.js
@@ -1,14 +1,8 @@
-const { ObjectID } = require('mongodb')
+const { ObjectId } = require('mongodb')
 const CoachingCompaniesModel = require('./CoachingCompanies.model')
 const mongoose = require('mongoose');
 const fs = require('fs')
 const path = require('path');
-const { tryEach } = require('async');
-const { success } = require('../paytm/paytm.controller');
-// const moment = require('moment')
-// const moment_timezone_1 = require("moment-timezone");
-// const serviceProductsModel = require('../serviceProducts/serviceProducts.model')
-// const ServiceAppointmentModel = require('../serviceAppointment/serviceAppointment.model');
 const jwt = require('jsonwebtoken')
 const bcrypt = require('bcryptjs')
 
@@ -39,9 +33,11 @@ module.exports = {
             }
             console.log(req.body, 'body')
             if (!CourseCompanyName || !CompanyOwnerName || !companyId || !HeadCourceCatId || !SubCourceCatId || !Street || !City || !State || !Country || !PostalCode || !Email || !Phone || !PanCardNo || !GstNo || !googleLocation || !Password || !ProviderType) {
-                const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-                if (fs.existsSync(newImagePath)) {
-                    fs.unlinkSync(newImagePath);
+                if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
                 }
 
                 return resp.status(400).json({ message: 'Please fill in all required fields', success: false });
@@ -50,10 +46,12 @@ module.exports = {
                 ConnectedWith.forEach((EachData) => {
                     if (EachData.connectedType) {
                         if (EachData.connectedIds.length == 0) {
-                            const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-                            if (fs.existsSync(newImagePath)) {
-                                fs.unlinkSync(newImagePath);
-                            }
+                           if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
+                }
 
                             return resp.status(400).json({ message: 'if you connected with someone then select them', success: false });
                         }
@@ -115,9 +113,11 @@ module.exports = {
 
 
             if (!result) {
-                const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-                if (fs.existsSync(newImagePath)) {
-                    fs.unlinkSync(newImagePath);
+               if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
                 }
 
                 return resp.status(400).json({ message: 'Something went wrong while saving the company', success: false });
@@ -125,10 +125,12 @@ module.exports = {
 
             return resp.status(200).json({ data: result, success: true });
         } catch (error) {
-            const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-            if (fs.existsSync(newImagePath)) {
-                fs.unlinkSync(newImagePath);
-            }
+           if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
+                }
             return resp.status(500).json({ error: error.message, success: false });
         }
     },
@@ -248,7 +250,7 @@ module.exports = {
                     _id: "$_id",
                     HeadCoachingCategories: { $first: "$HeadCoachingCategories" },
                     SubCoachingCategories: { $first: "$SubCoachingCategories" },
-                    ProviderType:{$first:"$ProviderType"},
+                    ProviderType: { $first: "$ProviderType" },
                     SkillsInfo: { $first: "$SkillsInfo" },
                     ProviderTypeInfo: { $first: "$ProviderTypeInfo" },
                     ConnectedInfos: {
@@ -494,10 +496,12 @@ module.exports = {
                         if (EachData.connectedIds.length == 0) {
                             console.log(ConnectedWith, 'hello');
 
-                            const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-                            if (fs.existsSync(newImagePath)) {
-                                fs.unlinkSync(newImagePath);
-                            }
+                            if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
+                }
 
                             return resp.status(400).json({ message: 'if you connected with someone then select them', success: false });
                         }
@@ -519,9 +523,11 @@ module.exports = {
             }
 
             if (!CoachingCompanyId || !HeadCourceCatId || !SubCourceCatId || !CourseCompanyName || !CompanyOwnerName || !companyId || !Street || !City || !State || !Country || !PostalCode || !Email || !Phone || !PanCardNo || !GstNo || !googleLocation) {
-                const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-                if (fs.existsSync(newImagePath)) {
-                    fs.unlinkSync(newImagePath);
+                if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
                 }
                 return resp.status(400).send('Please insert valid data');
 
@@ -579,10 +585,12 @@ module.exports = {
 
             }
         } catch (error) {
-            const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
-            if (fs.existsSync(newImagePath)) {
-                fs.unlinkSync(newImagePath);
-            }
+           if (req.file?.filename) {
+                    const newImagePath = path.join(__dirname, '..', '..', 'public', 'CoachingCompanyImage', req.file.filename);
+                    if (fs.existsSync(newImagePath)) {
+                        fs.unlinkSync(newImagePath);
+                    }
+                }
             return resp.status(400).json({ error: error.message, success: false });
         }
     },
